Create a common	threading header to abstract over platform differences.
The first job of the file is to define PTHREAD_RECURSIVE_MUTEX_INITIALIZER_NP
on Mac OS X, which defines it without the _NP prefix.

A followup commit will also define POSIX barrier functions, which do not
come with Mac OS X.

The goal of this overall patch series is to get multithreading support on mac.